added testcase for util
